+++ MatrixSDK/MXEnumConstants.m
@@ -27,7 +27,7 @@ NSString *const kMXContentPrefixPath = @"_matrix/media/v1";
 /**
  Prefix used in path of antivirus server API requests.
  */
-NSString *const kMXAntivirusAPIPrefixPathUnstable = @"_matrix/media_proxy/unstable/";
+NSString *const kMXAntivirusAPIPrefixPathUnstable = @"_matrix/media_proxy/unstable";
 
 /**
  Membership definitions - String version